cleanup, fixes, and refactoring
