+++ src/frontend/src/App.js
@@ -1,23 +1,23 @@
 import React from 'react';
-import AuthenticationGuard from './components/hoc/AuthenticationGuard';
-import Home from './components/page/home/Home';
-import OrderList from './components/orders/page/OrderList'
-import { Header, Footer } from 'shared-components';
-import { useHistory } from 'react-router-dom';
 import {
   Switch,
-  Route
+  Route,
+  useHistory
 } from "react-router-dom";
+import AuthenticationGuard from './components/hoc/AuthenticationGuard';
+import Home from './components/page/home/Home';
+import Orders from './components/page/orders/Orders';
+import { Header, Footer } from 'shared-components';
+import Countries from './components/page/countries/Countries';
 
 function App() {
-
   const header = {
     name: "Oracle Modernization",
     history: useHistory(),
   };
 
   return (
-    <div>
+    <main>
       <Header header={header} />
       <Switch>
         <Route exact path="/">
@@ -27,11 +27,14 @@ function App() {
           <AuthenticationGuard />
         </Route>
         <Route exact path="/orders">
-          <OrderList />
+          <Orders />
+        </Route>
+        <Route exact path="/countries">
+          <Countries />
         </Route>
       </Switch>
       <Footer className="bcgov-footer" />
-    </div>
+    </main>
   );
 }
