+++ src/frontend/src/components/composite/simple-modal/SimpleModal.css
