--- src/frontend/src/components/orders/page/OrderList.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import Orders from '../list/Orders'
-
-export default function OrderList() {
-
-  return (
-    <main>
-     
-      <div className="page">
-        <div className="content col-md-12">
-          <Orders />
-        </div>
-      </div>
-    </main>
-  );
-}
