--- src/frontend/src/components/page/contacts/ContactList.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import Contacts from '../../contacts/Contacts';
-
-export default function ContactList() {
-
-  return (
-    <main>
-     
-      <div className="page">
-        <div className="content col-md-12">
-          <Contacts />
-        </div>
-      </div>
-    </main>
-  );
-}
